RDKVREFPLT-2138 : [REFPLTV-2290] Update DISTRO VERSION in log
Reason for change: DISTRO_VERSION is still set to "2.0", and used in the logs "Welcome to RDK (A Yocto Project based Distro) 2.0 (dunfell)!", DISTRO_VERSION should be updated to track the version of the RDKV distribution. So, updating RDK_RELEASE value which in turn updates the DISTRO_VERSION.
Test Procedure: Build and Verified, The log message has been updated to the current RDK_RELEASE.
Risks: None.
Signed-off-by: Sharnetha_SK@comcast.com

Change-Id: Id769e923283eb2078d0f6835d932533b879eaf29